Struct im_rc::hashmap::HashMap

pub struct HashMap<K, V, S = RandomState> { /* fields omitted */ }

An unordered map.

An immutable hash map using [hash array mapped tries] 1.

Most operations on this map are O(logx n) for a suitably high x that it should be nearly O(1) for most maps. Because of this, it’s a great choice for a generic map as long as you don’t mind that keys will need to implement Hash and Eq.

Map entries will have a predictable order based on the hasher being used. Unless otherwise specified, this will be the standard RandomState hasher.

Implementations

impl<K, V> HashMap<K, V, RandomState>

#[must_use]
pub fn new() -> Self

Construct an empty hash map.

impl<K, V> HashMap<K, V, RandomState> where
K: Hash + Eq + Clone,
V: Clone,

#[must_use]
pub fn unit(k: K, v: V) -> HashMap<K, V>

Construct a hash map with a single mapping.

Examples

let map = HashMap::unit(123, "onetwothree");
assert_eq!(
  map.get(&123),
  Some(&"onetwothree")
);

impl<K, V, S> HashMap<K, V, S>

#[must_use]
pub fn is_empty(&self) -> bool

Test whether a hash map is empty.

Time: O(1)

Examples

assert!(
  !hashmap!{1 => 2}.is_empty()
);
assert!(
  HashMap::<i32, i32>::new().is_empty()
);

#[must_use]
pub fn len(&self) -> usize

Get the size of a hash map.

Time: O(1)

Examples

assert_eq!(3, hashmap!{
  1 => 11,
  2 => 22,
  3 => 33
}.len());

pub fn ptr_eq(&self, other: &Self) -> bool

Test whether two maps refer to the same content in memory.

This is true if the two sides are references to the same map, or if the two maps refer to the same root node.

This would return true if you’re comparing a map to itself, or if you’re comparing a map to a fresh clone of itself.

Time: O(1)

#[must_use]
pub fn with_hasher<RS>(hasher: RS) -> Self where
Rc<S>: From<RS>,

Construct an empty hash map using the provided hasher.

#[must_use]
pub fn hasher(&self) -> &Rc<S>

Get a reference to the map’s BuildHasher.

#[must_use]
pub fn new_from<K1, V1>(&self) -> HashMap<K1, V1, S> where
K1: Hash + Eq + Clone,
V1: Clone,

Construct an empty hash map using the same hasher as the current hash map.

#[must_use]
pub fn iter(&self) -> Iter<'_, K, V>

Get an iterator over the key/value pairs of a hash map.

Please note that the order is consistent between maps using the same hasher, but no other ordering guarantee is offered. Items will not come out in insertion order or sort order. They will, however, come out in the same order every time for the same map.

#[must_use]
pub fn keys(&self) -> Keys<'_, K, V>

Get an iterator over a hash map’s keys.

Please note that the order is consistent between maps using the same hasher, but no other ordering guarantee is offered. Items will not come out in insertion order or sort order. They will, however, come out in the same order every time for the same map.

#[must_use]
pub fn values(&self) -> Values<'_, K, V>

Get an iterator over a hash map’s values.

Please note that the order is consistent between maps using the same hasher, but no other ordering guarantee is offered. Items will not come out in insertion order or sort order. They will, however, come out in the same order every time for the same map.

pub fn clear(&mut self)

Discard all elements from the map.

This leaves you with an empty map, and all elements that were previously inside it are dropped.

Time: O(n)

Examples

let mut map = hashmap![1=>1, 2=>2, 3=>3];
map.clear();
assert!(map.is_empty());

impl<K, V, S> HashMap<K, V, S> where
K: Hash + Eq,
S: BuildHasher,

#[must_use]
pub fn get<BK: ?Sized>(&self, key: &BK) -> Option<&V> where
BK: Hash + Eq,
K: Borrow<BK>,

Get the value for a key from a hash map.

Time: O(log n)

Examples

let map = hashmap!{123 => "lol"};
assert_eq!(
  map.get(&123),
  Some(&"lol")
);

#[must_use]
pub fn get_key_value<BK: ?Sized>(&self, key: &BK) -> Option<(&K, &V)> where
BK: Hash + Eq,
K: Borrow<BK>,

Get the key/value pair for a key from a hash map.

Time: O(log n)

Examples

let map = hashmap!{123 => "lol"};
assert_eq!(
  map.get_key_value(&123),
  Some((&123, &"lol"))
);

#[must_use]
pub fn contains_key<BK: ?Sized>(&self, k: &BK) -> bool where
BK: Hash + Eq,
K: Borrow<BK>,

Test for the presence of a key in a hash map.

Time: O(log n)

Examples

let map = hashmap!{123 => "lol"};
assert!(
  map.contains_key(&123)
);
assert!(
  !map.contains_key(&321)
);

#[must_use]
pub fn is_submap_by<B, RM, F>(&self, other: RM, cmp: F) -> bool where
F: FnMut(&V, &B) -> bool,
RM: Borrow<HashMap<K, B, S>>,

Test whether a map is a submap of another map, meaning that all keys in our map must also be in the other map, with the same values.

Use the provided function to decide whether values are equal.

Time: O(n log n)

#[must_use]
pub fn is_proper_submap_by<B, RM, F>(&self, other: RM, cmp: F) -> bool where
F: FnMut(&V, &B) -> bool,
RM: Borrow<HashMap<K, B, S>>,

Test whether a map is a proper submap of another map, meaning that all keys in our map must also be in the other map, with the same values. To be a proper submap, ours must also contain fewer keys than the other map.

Use the provided function to decide whether values are equal.

Time: O(n log n)

#[must_use]
pub fn is_submap<RM>(&self, other: RM) -> bool where
V: PartialEq,
RM: Borrow<Self>,

Test whether a map is a submap of another map, meaning that all keys in our map must also be in the other map, with the same values.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 2 => 2};
let map2 = hashmap!{1 => 1, 2 => 2, 3 => 3};
assert!(map1.is_submap(map2));

#[must_use]
pub fn is_proper_submap<RM>(&self, other: RM) -> bool where
V: PartialEq,
RM: Borrow<Self>,

Test whether a map is a proper submap of another map, meaning that all keys in our map must also be in the other map, with the same values. To be a proper submap, ours must also contain fewer keys than the other map.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 2 => 2};
let map2 = hashmap!{1 => 1, 2 => 2, 3 => 3};
assert!(map1.is_proper_submap(map2));

let map3 = hashmap!{1 => 1, 2 => 2};
let map4 = hashmap!{1 => 1, 2 => 2};
assert!(!map3.is_proper_submap(map4));

impl<K, V, S> HashMap<K, V, S> where
K: Hash + Eq + Clone,
V: Clone,
S: BuildHasher,

#[must_use]
pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Get a mutable iterator over the values of a hash map.

Please note that the order is consistent between maps using the same hasher, but no other ordering guarantee is offered. Items will not come out in insertion order or sort order. They will, however, come out in the same order every time for the same map.

#[must_use]
pub fn get_mut<BK: ?Sized>(&mut self, key: &BK) -> Option<&mut V> where
BK: Hash + Eq,
K: Borrow<BK>,

Get a mutable reference to the value for a key from a hash map.

Time: O(log n)

Examples

let mut map = hashmap!{123 => "lol"};
if let Some(value) = map.get_mut(&123) {
    *value = "omg";
}
assert_eq!(
  map.get(&123),
  Some(&"omg")
);

pub fn insert(&mut self, k: K, v: V) -> Option<V>

Insert a key/value mapping into a map.

If the map already has a mapping for the given key, the previous value is overwritten.

Time: O(log n)

Examples

let mut map = hashmap!{};
map.insert(123, "123");
map.insert(456, "456");
assert_eq!(
  map,
  hashmap!{123 => "123", 456 => "456"}
);

pub fn remove<BK: ?Sized>(&mut self, k: &BK) -> Option<V> where
BK: Hash + Eq,
K: Borrow<BK>,

Remove a key/value pair from a map, if it exists, and return the removed value.

This is a copy-on-write operation, so that the parts of the set’s structure which are shared with other sets will be safely copied before mutating.

Time: O(log n)

Examples

let mut map = hashmap!{123 => "123", 456 => "456"};
assert_eq!(Some("123"), map.remove(&123));
assert_eq!(Some("456"), map.remove(&456));
assert_eq!(None, map.remove(&789));
assert!(map.is_empty());

pub fn remove_with_key<BK: ?Sized>(&mut self, k: &BK) -> Option<(K, V)> where
BK: Hash + Eq,
K: Borrow<BK>,

Remove a key/value pair from a map, if it exists, and return the removed key and value.

Time: O(log n)

Examples

let mut map = hashmap!{123 => "123", 456 => "456"};
assert_eq!(Some((123, "123")), map.remove_with_key(&123));
assert_eq!(Some((456, "456")), map.remove_with_key(&456));
assert_eq!(None, map.remove_with_key(&789));
assert!(map.is_empty());

#[must_use]
pub fn entry(&mut self, key: K) -> Entry<'_, K, V, S>

Get the Entry for a key in the map for in-place manipulation.

Time: O(log n)

#[must_use]
pub fn update(&self, k: K, v: V) -> Self

Construct a new hash map by inserting a key/value mapping into a map.

If the map already has a mapping for the given key, the previous value is overwritten.

Time: O(log n)

Examples

let map = hashmap!{};
assert_eq!(
  map.update(123, "123"),
  hashmap!{123 => "123"}
);

#[must_use]
pub fn update_with<F>(&self, k: K, v: V, f: F) -> Self where
F: FnOnce(V, V) -> V,

Construct a new hash map by inserting a key/value mapping into a map.

If the map already has a mapping for the given key, we call the provided function with the old value and the new value, and insert the result as the new value.

Time: O(log n)

#[must_use]
pub fn update_with_key<F>(&self, k: K, v: V, f: F) -> Self where
F: FnOnce(&K, V, V) -> V,

Construct a new map by inserting a key/value mapping into a map.

If the map already has a mapping for the given key, we call the provided function with the key, the old value and the new value, and insert the result as the new value.

Time: O(log n)

#[must_use]
pub fn update_lookup_with_key<F>(&self, k: K, v: V, f: F) -> (Option<V>, Self) where
F: FnOnce(&K, &V, V) -> V,

Construct a new map by inserting a key/value mapping into a map, returning the old value for the key as well as the new map.

If the map already has a mapping for the given key, we call the provided function with the key, the old value and the new value, and insert the result as the new value.

Time: O(log n)

#[must_use]
pub fn alter<F>(&self, f: F, k: K) -> Self where
F: FnOnce(Option<V>) -> Option<V>,

Update the value for a given key by calling a function with the current value and overwriting it with the function’s return value.

The function gets an Option<V> and returns the same, so that it can decide to delete a mapping instead of updating the value, and decide what to do if the key isn’t in the map.

Time: O(log n)

#[must_use]
pub fn without<BK: ?Sized>(&self, k: &BK) -> Self where
BK: Hash + Eq,
K: Borrow<BK>,

Construct a new map without the given key.

Construct a map that’s a copy of the current map, absent the mapping for key if it’s present.

Time: O(log n)

pub fn retain<F>(&mut self, f: F) where
F: FnMut(&K, &V) -> bool,

Filter out values from a map which don’t satisfy a predicate.

This is slightly more efficient than filtering using an iterator, in that it doesn’t need to rehash the retained values, but it still needs to reconstruct the entire tree structure of the map.

Time: O(n log n)

Examples

let mut map = hashmap!{1 => 1, 2 => 2, 3 => 3};
map.retain(|k, v| *k > 1);
let expected = hashmap!{2 => 2, 3 => 3};
assert_eq!(expected, map);

#[must_use]
pub fn extract<BK: ?Sized>(&self, k: &BK) -> Option<(V, Self)> where
BK: Hash + Eq,
K: Borrow<BK>,

Remove a key/value pair from a map, if it exists, and return the removed value as well as the updated map.

Time: O(log n)

#[must_use]
pub fn extract_with_key<BK: ?Sized>(&self, k: &BK) -> Option<(K, V, Self)> where
BK: Hash + Eq,
K: Borrow<BK>,

Remove a key/value pair from a map, if it exists, and return the removed key and value as well as the updated list.

Time: O(log n)

#[must_use]
pub fn union(self, other: Self) -> Self

Construct the union of two maps, keeping the values in the current map when keys exist in both maps.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 3 => 3};
let map2 = hashmap!{2 => 2, 3 => 4};
let expected = hashmap!{1 => 1, 2 => 2, 3 => 3};
assert_eq!(expected, map1.union(map2));

#[must_use]
pub fn union_with<F>(self, other: Self, f: F) -> Self where
F: FnMut(V, V) -> V,

Construct the union of two maps, using a function to decide what to do with the value when a key is in both maps.

The function is called when a value exists in both maps, and receives the value from the current map as its first argument, and the value from the other map as the second. It should return the value to be inserted in the resulting map.

Time: O(n log n)

#[must_use]
pub fn union_with_key<F>(self, other: Self, f: F) -> Self where
F: FnMut(&K, V, V) -> V,

Construct the union of two maps, using a function to decide what to do with the value when a key is in both maps.

The function is called when a value exists in both maps, and receives a reference to the key as its first argument, the value from the current map as the second argument, and the value from the other map as the third argument. It should return the value to be inserted in the resulting map.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 3 => 4};
let map2 = hashmap!{2 => 2, 3 => 5};
let expected = hashmap!{1 => 1, 2 => 2, 3 => 9};
assert_eq!(expected, map1.union_with_key(
    map2,
    |key, left, right| left + right
));

#[must_use]
pub fn unions<I>(i: I) -> Self where
S: Default,
I: IntoIterator<Item = Self>,

Construct the union of a sequence of maps, selecting the value of the leftmost when a key appears in more than one map.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 3 => 3};
let map2 = hashmap!{2 => 2};
let expected = hashmap!{1 => 1, 2 => 2, 3 => 3};
assert_eq!(expected, HashMap::unions(vec![map1, map2]));

#[must_use]
pub fn unions_with<I, F>(i: I, f: F) -> Self where
S: Default,
I: IntoIterator<Item = Self>,
F: Fn(V, V) -> V,

Construct the union of a sequence of maps, using a function to decide what to do with the value when a key is in more than one map.

The function is called when a value exists in multiple maps, and receives the value from the current map as its first argument, and the value from the next map as the second. It should return the value to be inserted in the resulting map.

Time: O(n log n)

#[must_use]
pub fn unions_with_key<I, F>(i: I, f: F) -> Self where
S: Default,
I: IntoIterator<Item = Self>,
F: Fn(&K, V, V) -> V,

Construct the union of a sequence of maps, using a function to decide what to do with the value when a key is in more than one map.

The function is called when a value exists in multiple maps, and receives a reference to the key as its first argument, the value from the current map as the second argument, and the value from the next map as the third argument. It should return the value to be inserted in the resulting map.

Time: O(n log n)

#[must_use]
pub fn difference(self, other: Self) -> Self

Construct the symmetric difference between two maps by discarding keys which occur in both maps.

This is an alias for the symmetric_difference method.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 3 => 4};
let map2 = hashmap!{2 => 2, 3 => 5};
let expected = hashmap!{1 => 1, 2 => 2};
assert_eq!(expected, map1.difference(map2));

#[must_use]
pub fn symmetric_difference(self, other: Self) -> Self

Construct the symmetric difference between two maps by discarding keys which occur in both maps.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 3 => 4};
let map2 = hashmap!{2 => 2, 3 => 5};
let expected = hashmap!{1 => 1, 2 => 2};
assert_eq!(expected, map1.symmetric_difference(map2));

#[must_use]
pub fn difference_with<F>(self, other: Self, f: F) -> Self where
F: FnMut(V, V) -> Option<V>,

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both.

This is an alias for the symmetric_difference_with method.

Time: O(n log n)

#[must_use]
pub fn symmetric_difference_with<F>(self, other: Self, f: F) -> Self where
F: FnMut(V, V) -> Option<V>,

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both.

Time: O(n log n)

#[must_use]
pub fn difference_with_key<F>(self, other: Self, f: F) -> Self where
F: FnMut(&K, V, V) -> Option<V>,

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both. The function receives the key as well as both values.

This is an alias for the symmetric_difference_with_key method.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 3 => 4};
let map2 = hashmap!{2 => 2, 3 => 5};
let expected = hashmap!{1 => 1, 2 => 2, 3 => 9};
assert_eq!(expected, map1.difference_with_key(
    map2,
    |key, left, right| Some(left + right)
));

#[must_use]
pub fn symmetric_difference_with_key<F>(self, other: Self, f: F) -> Self where
F: FnMut(&K, V, V) -> Option<V>,

Construct the symmetric difference between two maps by using a function to decide what to do if a key occurs in both. The function receives the key as well as both values.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 3 => 4};
let map2 = hashmap!{2 => 2, 3 => 5};
let expected = hashmap!{1 => 1, 2 => 2, 3 => 9};
assert_eq!(expected, map1.symmetric_difference_with_key(
    map2,
    |key, left, right| Some(left + right)
));

#[must_use]
pub fn relative_complement(self, other: Self) -> Self

Construct the relative complement between two maps by discarding keys which occur in other.

Time: O(m log n) where m is the size of the other map

Examples

let map1 = ordmap!{1 => 1, 3 => 4};
let map2 = ordmap!{2 => 2, 3 => 5};
let expected = ordmap!{1 => 1};
assert_eq!(expected, map1.relative_complement(map2));

#[must_use]
pub fn intersection(self, other: Self) -> Self

Construct the intersection of two maps, keeping the values from the current map.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 2 => 2};
let map2 = hashmap!{2 => 3, 3 => 4};
let expected = hashmap!{2 => 2};
assert_eq!(expected, map1.intersection(map2));

#[must_use]
pub fn intersection_with<B, C, F>(
self,
other: HashMap<K, B, S>,
f: F
) -> HashMap<K, C, S> where
B: Clone,
C: Clone,
F: FnMut(V, B) -> C,

Construct the intersection of two maps, calling a function with both values for each key and using the result as the value for the key.

Time: O(n log n)

#[must_use]
pub fn intersection_with_key<B, C, F>(
self,
other: HashMap<K, B, S>,
f: F
) -> HashMap<K, C, S> where
B: Clone,
C: Clone,
F: FnMut(&K, V, B) -> C,

Construct the intersection of two maps, calling a function with the key and both values for each key and using the result as the value for the key.

Time: O(n log n)

Examples

let map1 = hashmap!{1 => 1, 2 => 2};
let map2 = hashmap!{2 => 3, 3 => 4};
let expected = hashmap!{2 => 5};
assert_eq!(expected, map1.intersection_with_key(
    map2,
    |key, left, right| left + right
));