Module im_rc::hashset

An unordered set.

An immutable hash set using [hash array mapped tries] 1.

Most operations on this set are O(logx n) for a suitably high x that it should be nearly O(1) for most sets. Because of this, it’s a great choice for a generic set as long as you don’t mind that values will need to implement Hash and Eq.

Values will have a predictable order based on the hasher being used. Unless otherwise specified, this will be the standard RandomState hasher.

Structs

ConsumingIter
A consuming iterator over the elements of a set.

HashSet
An unordered set.

HashSetPool
A memory pool for the appropriate node type.

Iter
An iterator over the elements of a set.